use float64 for orderindexes

+++ clickup/checklists.go
@@ -24,19 +24,19 @@ type ChecklistItemRequest struct {
 }
 
 type Checklist struct {
-	ID         string `json:"id"`
-	TaskID     string `json:"task_id"`
-	Name       string `json:"name"`
-	Orderindex int    `json:"orderindex"`
-	Resolved   int    `json:"resolved"`
-	Unresolved int    `json:"unresolved"`
-	Items      []Item `json:"items,omitempty"`
+	ID         string  `json:"id"`
+	TaskID     string  `json:"task_id"`
+	Name       string  `json:"name"`
+	Orderindex float64 `json:"orderindex"`
+	Resolved   int     `json:"resolved"`
+	Unresolved int     `json:"unresolved"`
+	Items      []Item  `json:"items,omitempty"`
 }
 
 type Item struct {
 	ID          string        `json:"id"`
 	Name        string        `json:"name"`
-	Orderindex  int           `json:"orderindex"`
+	Orderindex  float64       `json:"orderindex"`
 	Assignee    User          `json:"assignee"`
 	Resolved    bool          `json:"resolved"`
 	Parent      interface{}   `json:"parent"`
